@@ -41,10 +41,10 @@
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.8",
 		"prettier-plugin-tailwindcss": "^0.6.8",
-		"svelte": "^5.1.16",
+		"svelte": "^5.2.0",
 		"svelte-body": "^2.0.0",
 		"svelte-check": "^4.0.7",
-		"tailwindcss": "^3.4.14",
+		"tailwindcss": "^3.4.15",
 		"tslib": "^2.8.1",
 		"typescript": "^5.6.3",
 		"vite": "^5.4.11"
