@@ -64,7 +64,7 @@
     "prettier-plugin-svelte": "^3.2.8",
     "prettier-plugin-tailwindcss": "^0.6.8",
     "randomcolor": "^0.6.2",
-    "svelte": "^5.1.16",
+    "svelte": "^5.2.0",
     "svelte-check": "^4.0.7",
     "svelte-echarts": "1.0.0-rc3",
     "svelte-floating-ui": "^1.5.9",
@@ -72,7 +72,7 @@
     "svelte-selecto": "^1.26.3",
     "svelte-use-debounce": "^0.0.2",
     "svelte-virtual-scroll-list": "^1.3.0",
-    "tailwindcss": "^3.4.14",
+    "tailwindcss": "^3.4.15",
     "tiny-glob": "^0.2.9",
     "tslib": "^2.8.1",
     "tslog": "^4.9.3",
